% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FractionalRidgeMain.R
\name{frac_ridge_pred}
\alias{frac_ridge_pred}
\title{A function to do the prediction using X (design matrix) and beta_ridge (estimated coefficients)}
\usage{
frac_ridge_pred(X, beta_ridge, intercept = FALSE)
}
\arguments{
\item{X}{A size n*p matrix (design matrix for regression, with n number of observations and p number of variables)}

\item{beta_ridge}{A size (p, f, t) array (f - the number of fractions in previous frac_ridge input) obtained from frac_ridge(...)$coefs}

\item{intercept}{A logical argument indicates whether the user needs an intercept or not}
}
\value{
A size (p, t, f) array with each array\link{,,i} representing the estimated y for each user-defined fraction in frac_ridge(...)
}
\description{
A function to do the prediction using X (design matrix) and beta_ridge (estimated coefficients)
}
\examples{
x = matrix(rnorm(100*10), 100, 10)
y = rnorm(100)
beta_ridge = frac_ridge(x, y, 0.1, intercept = TRUE, scale_x = FALSE)$coefs
y_pred = frac_ridge_pred(x, beta_ridge, TRUE)
}
