% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FractionalRidgeMain.R
\name{r2_score}
\alias{r2_score}
\title{A function to calculate Rsquared score using frac_ridge output}
\usage{
r2_score(y, X, beta_ridge, intercept = FALSE)
}
\arguments{
\item{y}{A size n*t matrix (data matrix, with n number of observations and t number of targets)}

\item{X}{A size n*p matrix (design matrix for regression, with n number of observations and p number of variables)}

\item{beta_ridge}{A size (p, f, t) array (f - the number of fractions in previous frac_ridge input) obtained from frac_ridge(...)$coefs}

\item{intercept}{A logical argument indicates whether the user needs an intercept or not}
}
\value{
r2_score A size f vector consisting of average Rsquared score for each fraction
}
\description{
A function to calculate Rsquared score using frac_ridge output
}
\examples{
x = matrix(rnorm(100*10), 100, 10)
y = rnorm(100)
beta_ridge = frac_ridge(x, y, 0.1, intercept = TRUE, scale_x = FALSE)$coefs
r2 = r2_score(y, x, beta_ridge, TRUE)
}
