% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scaleMat}
\alias{scaleMat}
\title{A fast implementation of scaling the matrix using meanVec and stdVec defined in rcpp}
\usage{
scaleMat(x)
}
\arguments{
\item{x}{A size n*p matrix}
}
\value{
scaled matrix x
}
\description{
A fast implementation of scaling the matrix using meanVec and stdVec defined in rcpp
}
