% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FractionalRidgeMain.R
\name{svd_x}
\alias{svd_x}
\title{A singular value decomposition of matrix}
\usage{
svd_x(X, y)
}
\arguments{
\item{X}{A size n*p matrix (design matrix for regression, with n number of observations and p number of variables)}

\item{y}{A size n*t matrix (data matrix, with n number of observations and t number of targets)}
}
\value{
A list containing singular values of X (lambda_val), unitary matrix vt from SVD (vt), OLS coefficients in the rotated space (beta_ols), and y in the rotated space (y_trsf)
}
\description{
A singular value decomposition of matrix
}
\examples{
x = matrix(rnorm(100*10), 100, 10)
y = rnorm(100)
res = svd_x(x, y)
}
